♻️ refactor : HeroSection 컴포넌트 추가 및 메인 페이지 구조 개선 #1
